# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Houston Mobile Mechanic API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stripe/checkout:
    post:
      operationId: createCheckout
      tags: [shop]
      summary: Create a Stripe Checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
...
  /stripe/deposit:
    post:
      operationId: createDeposit
      tags: [shop]
      summary: Create a $50 service booking deposit checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
...
  /technician-applications:
    post:
      operationId: submitTechnicianApplication
      tags: [applications]
      summary: Submit a public technician application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianApplicationSubmission"
    get:
      operationId: listTechnicianApplications
      tags: [applications]
      summary: List technician applications for owner review
      security: [{ clerk: [] }]
      responses:
        "200":
          description: Technician applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechnicianApplication"
        "401":
          description: Authentication required
        "403":
          description: Owner or admin access required
  /technician-applications/{id}:
    get:
      operationId: getTechnicianApplication
      tags: [applications]
      summary: Read one technician application for owner review
      security: [{ clerk: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianApplication"
        "401":
          description: Authentication required
        "403":
          description: Owner or admin access required
        "404":
          description: Application not found
    patch:
      operationId: updateTechnicianApplication
      tags: [applications]
      summary: Update a technician application review
      security: [{ clerk: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianApplicationUpdate"
      responses:
        "200":
          description: Updated technician application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianApplication"
        "400":
          description: Invalid review update
        "401":
          description: Authentication required
        "403":
          description: Owner or admin access required
        "404":
          description: Application not found
...
components:
  schemas:
...
    CheckoutInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [priceId, quantity]
            properties:
              priceId:
                type: string
              quantity:
                type: integer
        customerEmail:
          type: ["string", "null"]
    CheckoutResult:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    DepositInput:
      type: object
      properties:
        customerEmail:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
...
    TechnicianApplication:
      type: object
      required: [id, fullName, phone, email, serviceArea, experience, specialties, availability, introduction, status, ownerNotes, reviewedAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        serviceArea: { type: string }
        experience: { type: string }
        specialties: { type: string }
        availability: { type: string }
        introduction: { type: ["string", "null"] }
        status: { type: string, enum: [new, reviewing, approved, declined] }
        ownerNotes: { type: ["string", "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TechnicianApplicationInput:
      type: object
      required: [fullName, phone, email, serviceArea, experience, specialties, availability]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 120 }
        phone: { type: string, minLength: 7, maxLength: 40 }
        email: { type: string, format: email, maxLength: 254 }
        serviceArea: { type: string, minLength: 1, maxLength: 160 }
        experience: { type: string, minLength: 1, maxLength: 120 }
        specialties: { type: string, minLength: 1, maxLength: 500 }
        availability: { type: string, minLength: 1, maxLength: 120 }
        introduction: { type: ["string", "null"], maxLength: 2000 }
    TechnicianApplicationSubmission:
      type: object
      required: [submitted, message]
      properties:
        submitted: { type: boolean }
        message: { type: string }
    TechnicianApplicationUpdate:
      type: object
      properties:
        status: { type: string, enum: [new, reviewing, approved, declined] }
        ownerNotes: { type: ["string", "null"], maxLength: 4000 }
...
  securitySchemes:
    clerk:
      type: http
      scheme: bearer